% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-context.R
\name{cluster_context}
\alias{cluster_context}
\title{Create the cluster context}
\usage{
cluster_context(x, y)
}
\arguments{
\item{x}{The input data (numeric matrix).}

\item{y}{Cluster membership indices (integer vector). -1 indicates noise.}
}
\value{
A list with the following elements:
\itemize{
\item \code{k}: The number of clusters (integer).
\item \code{centroids}: Centroid coordinates (k x n matrix).
\item \code{mass}: Cluster sizes (integer vector of length k).
\item \code{mn_da}: Mean Euclidean distances of the cluster members to
their centroid (double vector of length k).
\item \code{md_da}: Median Euclidean distances of the cluster members to
their centroid (double vector of length k).
\item \code{sd_da}: Standard deviation of the Euclidean distances of the
cluster members to their centroid (double vector of length k).
\item \code{de}: Euclidean distance matrix for the centroids (k x k
matrix).
\item \code{outliers}: The number of outliers (integer).
}
}
\description{
Create the cluster context
}
\examples{
# TODO
}
