% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-symbol.R
\name{draw_symbol}
\alias{draw_symbol}
\title{Draw ideogram for clustering solution}
\usage{
draw_symbol(cc, gv, rc, similar_density_th = 3, radii_ratio_th = 2)
}
\arguments{
\item{cc}{The \code{\link[cluster_context]{cluster context}}.}

\item{gv}{The \code{\link[gval]{GOI cluster validity}} indices object.}

\item{rc}{The \code{\link[refinement_context]{refinement context}}.}

\item{similar_density_th}{Threshold of the (relative) difference between the
cluster with the highest density and the cluster with the lowest density for
the clustering to exhibit inter-cluster density differences.}

\item{radii_ratio_th}{Threshold of the extended-to-core ratio for the
classification as "long-tailed" clusters.}
}
\description{
Draw ideogram for clustering solution
}
\examples{
# TODO
}
