% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gval.R
\name{gval}
\alias{gval}
\title{Compute GOI cluster validity indices}
\usage{
gval(cc)
}
\arguments{
\item{cc}{The \code{\link[cluster_context]{cluster context}}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{g_str}: Strict G-index (double).
\item \code{g_rex}: Relaxed G-index (double).
\item \code{g_min}: Min G-index (double)-
\item \code{oi_st}: Cluster-individual strict overlap indices (double
vector of length k).
\item \code{oi_rx}: Cluster-individual relaxed overlap indices (double
vector of length k).
\item \code{oi_mn}: Cluster-individual min overlap indices (double vector
of length k).
\item \code{ext_r}: Extended cluster radii (double vector of length k).
\item \code{str_r}: Strict cluster radii (double vector of length k).
\item \code{vol_r}: Extended-to-core ratio.
}
}
\description{
Compute GOI cluster validity indices
}
\examples{
# TODO
}
