% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.R
\name{refine}
\alias{refine}
\title{Refine a clustering solution using the CluReAL algorithm}
\usage{
refine(
  x,
  y,
  cc,
  gv,
  rc,
  rep = 0,
  min_rdens = -0.8,
  min_mass = 0.005,
  out_sens = 0.75,
  prun_level = 0
)
}
\arguments{
\item{x}{The input data (numeric matrix).}

\item{y}{Cluster membership indices (integer vector). -1 indicates noise.}

\item{cc}{The \code{\link[cluster_context]{cluster context}}.}

\item{gv}{The \code{\link[gval]{GOI cluster validity}} indices object.}

\item{rc}{The \code{\link[refinement_context]{refinement context}}.}

\item{rep}{The number of refinement repetitions.}

\item{min_rdens}{The minimum density threshold (relative to the overall
density) of a cluster to not be considered noise.}

\item{min_mass}{The relative minimum size of a cluster to not be considered
noise.}

\item{out_sens}{Outlier sensitivity. Specifies how far away outliers must be
from the centroids to avoid trying to merge them with the clusters.}

\item{prun_level}{Pruning level. 0: normal (default),
1: high cluster overlap expected,
2: very high cluster overlap expected}
}
\value{
A list with two elements:
\itemize{
\item \code{y}: Vector of the refined cluster membership indices.
\item \code{cc}: Refined cluster context
}
}
\description{
Run the CluReAL V2 algorithm on a data matrix and a vector of cluster
membership indices.
The output is a list with the refined cluster membership indices and the
the cluster context.
}
\details{
For a description of the algorithm, see:

Iglesias, Felix, Tanja Zseby, and Arthur Zimek. "Clustering refinement."
International Journal of Data Science and Analytics (2021): 1-21.
URL: \url{https://doi.org/10.1007/s41060-021-00275-z}.
}
\examples{
# TODO
}
